+++ index.py
@@ -14,40 +14,35 @@ app = Flask(__name__)
 
 app.secret_key = '1234'
 
+app.config.update(SESSION_COOKIE_SAMESITE="None", SESSION_COOKIE_SECURE=True)
+
 site = Blueprint('site', __name__, template_folder='templates')
 
 database = Database()
-logged_in = False
-
-logger.info('passou')
 
 @app.route('/login', methods=['GET', 'POST'])
 def login():
 
-    session['logged_in'] = False
-
-    if request.method == 'POST' and 'username' in request.form and 'password' in request.form:
+    if request.method == 'POST' and 'email' in request.form and 'password' in request.form:
         
         # Create variables for easy access
         username = request.form['email']
         password = request.form['password']
 
         if database.user_exists(username, password):
+
+            logger.info("Login valido")
             session['logged_in'] = True
-            logged_in = session['logged_in']
-            return redirect(url_for('home'))
+            return redirect(url_for(f'home', user_id = database.get_user_id(username, password)))
             
         else:
-            flash('login inválido')
+            logger.info("Login invalido")
     
     return render_template('login.html')
  
-@app.route("/home")
-def home():
-    if logged_in:
-        return render_template('home.html')
-    else:
-        return redirect(url_for('login'))
+@app.route("/home/<user_id>")
+def home(user_id):
+        return render_template('home.html', user_id = user_id)
 
 @app.route("/register", methods=['GET','POST'])
 def register():
@@ -94,20 +89,20 @@ def ajustaComparativo(base, ref):
 
     return resp
 
-@app.route('/busca/<userid>', methods=['GET', 'POST'])
-def busca(userid):
-    conn = sqlite3.connect('database.db')
+@app.route('/busca/<user_id>', methods=['GET', 'POST'])
+def busca(user_id):
+    #todo: checar se usuário esta logado
+    conn = sqlite3.connect('minhafacul.db')
     c = conn.cursor()
 
     if request.method == 'POST':
         # Atualiza os dados de local e curso selecionado
-        #TODO
         conn.execute("UPDATE USUARIO SET CURSO_ID=?, LOCAL_TXT=?, LOCAL_LAT=?, LOCAL_LON=? WHERE USUARIO_ID=?;",
-                     (request.form['curso_id'], request.form['local_txt'], request.form['local_lat'], request.form['local_lon'], userid))
+                     (request.form['curso_id'], request.form['local_txt'], request.form['local_lat'], request.form['local_lon'], user_id))
         conn.commit()
 
     # Consulda dados do usário para construir a busca
-    c.execute("SELECT USUARIO_ID, EMAIL, CURSO_ID, LOCAL_TXT, LOCAL_LAT, LOCAL_LON FROM USUARIO WHERE USUARIO_ID=?", (userid))
+    c.execute("SELECT USUARIO_ID, EMAIL, CURSO_ID, LOCAL_TXT, LOCAL_LAT, LOCAL_LON FROM USUARIO WHERE USUARIO_ID=?", (user_id, ))
     userdat = c.fetchall()
 
     # Consulta a tabela de cursos e retorna uma lista de tuplas com os dados
@@ -116,13 +111,130 @@ def busca(userid):
 
     if (userdat[0][2]>0 and userdat[0][4] != '' and userdat[0][5] != ''):
         # Curso já selecionado e local definido
-        c.execute("SELECT * FROM COMPARATIVO WHERE CURSO_ID=?", (userdat[0][2]))
+        c.execute(f"SELECT * FROM COMPARATIVO WHERE CURSO_ID={userdat[0][2]}")
         comparativo = c.fetchall()
         comparativo = ajustaComparativo(comparativo, (userdat[0][4], userdat[0][5]))
     else:
         comparativo = None
 
     return render_template('busca.html', comparativo=comparativo, userdat=userdat, cursos=cursos)
+
+
+@app.route('/admin/<blk>', methods=['GET', 'POST'])
+def admin(blk):
+    #TODO: checar se o usuário está logado e é administrador!
+
+    conn = sqlite3.connect('minhafacul.db')
+    c = conn.cursor()
+
+    c.execute("SELECT CURSO_ID, CURSO FROM CURSO ORDER BY CURSO")
+    cursos = c.fetchall()
+
+    c.execute("SELECT FACULDADE_ID, FACULADE, LOCAL_TXT, LOCAL_LAT, LOCAL_LON FROM FACULDADE ORDER BY FACULADE")
+    faculdades = c.fetchall()
+
+    c.execute("SELECT HISTORICO_ID, FACULADE, CURSO, ANO, CANDIDATOS, VAGAS FROM HISTORICO_DET ORDER BY ANO, FACULADE, CURSO")
+    historico = c.fetchall()
+
+    c.execute("SELECT USUARIO_ID, EMAIL, ADMINISTRADOR FROM USUARIO ORDER BY EMAIL")
+    usuarios = c.fetchall()
+
+    return render_template('admin.html', cursos=cursos, faculdades=faculdades, usuarios=usuarios, historico=historico, blk=blk)
+
+@app.route('/curso/I', methods=['GET', 'POST'])
+def curso_add():
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("INSERT INTO CURSO (CURSO) VALUES (?)", (request.form['curson'], ))
+    conn.commit()
+    return redirect(url_for('admin', blk=1))
+
+@app.route('/curso/D/<id>', methods=['GET', 'POST'])
+def curso_del(id):
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("DELETE FROM CURSO WHERE CURSO_ID=?", (id, ))
+    conn.commit()
+    return redirect(url_for('admin', blk=1))
+
+
+@app.route('/curso/U/<id>', methods=['GET', 'POST'])
+def curso_upd(id):
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("UPDATE CURSO SET CURSO=? WHERE CURSO_ID=?", (request.form['curson'], id))
+    conn.commit()
+    return redirect(url_for('admin', blk=1))
+
+
+@app.route('/faculdade/I', methods=['GET', 'POST'])
+def faculdade_add():
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("INSERT INTO FACULDADE (FACULADE, LOCAL_TXT, LOCAL_LAT, LOCAL_LON) VALUES(?, ?, ?, ?)",
+                 (request.form['faculaden'], request.form['local_txt'], request.form['local_lat'], request.form['local_lon']))
+    conn.commit()
+    return redirect(url_for('admin', blk=2))
+
+
+@app.route('/faculdade/D/<id>', methods=['GET', 'POST'])
+def faculdade_del(id):
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("DELETE FROM FACULDADE WHERE FACULDADE_ID=?", (id, ))
+    conn.commit()
+    return redirect(url_for('admin', blk=2))
+
+
+@app.route('/faculdade/U/<id>', methods=['GET', 'POST'])
+def faculdade_upd(id):
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("UPDATE FACULDADE SET FACULADE=?, LOCAL_TXT=?, LOCAL_LAT=?, LOCAL_LON=? WHERE FACULDADE_ID=?",
+                 (request.form['faculaden'], request.form['local_txt'], request.form['local_lat'], request.form['local_lon'], id))
+    conn.commit()
+    return redirect(url_for('admin', blk=2))
+
+
+@app.route('/historico/D/<id>', methods=['GET', 'POST'])
+def historico_del(id):
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("DELETE FROM HISTORICO WHERE HISTORICO_ID=?", (id, ))
+    conn.commit()
+    return redirect(url_for('admin', blk=3))
+
+
+@app.route('/historico/U/<id>', methods=['GET', 'POST'])
+def historico_upd(id):
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("UPDATE HISTORICO SET ANO=?, CANDIDATOS=?, VAGAS=? WHERE HISTORICO_ID=?",
+                 (request.form['ano'], request.form['candidatos'], request.form['vagas'], id))
+    conn.commit()
+    return redirect(url_for('admin', blk=3))
+
+
+@app.route('/usuario/D/<id>', methods=['GET', 'POST'])
+def usuario_del(id):
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("DELETE FROM USUARIO WHERE USUARIO_ID=?", (id, ))
+    conn.commit()
+    return redirect(url_for('admin', blk=4))
+
+
+@app.route('/usuario/U/<id>', methods=['GET', 'POST'])
+def usuario_upd(id):
+    conn = sqlite3.connect('minhafacul.db')
+    conn.execute("UPDATE USUARIO SET ADMINISTRADOR=1-ADMINISTRADOR WHERE USUARIO_ID=?", (id, ))
+    conn.commit()
+    return redirect(url_for('admin', blk=4))
+
+@app.route('/profile/<user_id>', methods=['GET', 'POST'])
+def profile(user_id):
+
+    if (request.method == 'POST' and 'password' in request.form and 'password_c' in request.form):
+        
+        password = request.form['password']
+        database.alter_password(user_id, password)
+
+    email = database.get_user_email(user_id)
+    local = database.get_user_local(user_id)
+
+    return render_template('profile.html', email = email, local = local)
+
 
 if __name__ == '__main__':
     app.run(debug=True)
