+++ templates/home.html
@@ -4,9 +4,14 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>Home</title>
 </head>
 <body>
-    <hr> HOME 
+    <hr><h1>HOME</h1><hr>
+    <br/><br/>
+    <a href="/profile/{{user_id}}" type="button">Perfil do Usuário</a> 
+    <br/><br/>
+    <a href="/busca/{{user_id}}" type="button">Realizar Busca</a><br/>
+
 </body>
 </html>
