+++ templates/login.html
@@ -11,9 +11,10 @@
         <div class="tema">
         <div class="int">
         <form method="POST">
+            
         <br/>Email:
-
         <br/><input type="text" name="email"/> 
+
         <br/>Senha:
         <br/><input type="password" name="password"/> <br/>
